big refactor pt 2
